Update-drizzle-files

+++ next-js-chat/drizzle/0000_neat_xorn.sql
@@ -1,13 +1,13 @@
 CREATE TABLE IF NOT EXISTS "fetched_messages" (
 	"id" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name "fetched_messages_id_seq" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),
-	"timestamp" integer NOT NULL,
+	"timestamp" bigint NOT NULL,
 	"user_id" integer NOT NULL
 );
 --> statement-breakpoint
 CREATE TABLE IF NOT EXISTS "messages" (
 	"id" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name "messages_id_seq" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),
 	"message" varchar(255) NOT NULL,
-	"timestamp" integer NOT NULL,
+	"timestamp" bigint NOT NULL,
 	"user_id" integer NOT NULL
 );
 --> statement-breakpoint
